Done delete by correct user

+++ app/__init__.py
@@ -102,16 +102,17 @@ def show_one_thing(id):
 def add_a_thing():
     # Get the data from the form
     name  = request.form.get("name")
-    price = request.form.get("price")
 
     # Sanitise the inputs
     name = html.escape(name)
-    price = html.escape(price)
+
+    # Get the user_id from the session
+    user_id = session["user_id"]
 
     with connect_db() as client:
         # Add the thing to the DB
-        sql = "INSERT INTO things (name, price) VALUES (?, ?)"
-        values = [name, price]
+        sql = "INSERT INTO things (name, user_id) VALUES (?, ?)"
+        values = [name, user_id]
         client.execute(sql, values)
 
         # Go back to the home page
@@ -191,8 +192,12 @@ def login_user():
 @app.get("/delete/<int:id>")
 def delete_a_thing(id):
     with connect_db() as client:
-        # Delete the thing from the DB
-        sql = "DELETE FROM things WHERE id=?"
+        # Get our user id from session
+        user_id = session["user_id"]
+
+        # Delete the thing from the DB 
+        # Checking if we are the owner
+        sql = "DELETE FROM things WHERE id=? AND user_id=?"
         values = [id]
         client.execute(sql, values)
 
@@ -200,4 +205,12 @@ def delete_a_thing(id):
         flash("Thing deleted", "warning")
         return redirect("/things")
 
-
+#-----------------------------------------------------------
+# Route for logging out a user
+#-----------------------------------------------------------
+@app.get("/logout")
+def logout():
+    session.pop("user_id")
+    session.pop("user_name")
+    flash("You have been logged out", "success")
+    return redirect("/")
